moving openReplay to utils
