+++ apps/web/pages/home.js
@@ -40,7 +40,8 @@ import {
 import { useAuthChannel, useLogout } from '../src/hooks';
 import { txType, getActivityId } from '../src/utils/transaction';
 import { Routes } from '../src/config';
-import { EVENTS, logEvent, openReplayTracker } from '../src/utils/analytics';
+import { EVENTS, logEvent } from '../src/utils/analytics';
+import { openReplayTracker } from '../src/utils/openReplay';
 
 const tabs = {
   EXPLORE: 0,
