+++ apps/web/src/utils/analytics.js
@@ -1,5 +1,4 @@
 import amplitude from 'amplitude-js';
-import Tracker from '@openreplay/tracker/cjs';
 import { App } from '../config';
 import { isClient } from './environment';
 
@@ -64,8 +63,3 @@ export const EVENTS = {
 export const logEvent = (event) => {
   amplitude.getInstance().logEvent(event);
 };
-
-export const openReplayTracker = new Tracker({
-  projectKey: App.openReplay.projectKey,
-  // __DISABLE_SECURE_MODE: true, // enable this during development mode
-});
